Put IR class in a separate file.
