+++ src/LinAlg/CMakeLists.txt
@@ -99,6 +99,7 @@ set(hiopLinAlg_PARDISO_SRC
 
 set(hiopLinAlg_CUSOLVER_LU_SRC
   hiopLinSolverSparseCUSOLVER.cpp
+  IterativeRefinement.cpp
 )
 
 set(hiopLinAlg_CUSOLVER_CHOL_SRC
